+++ database/migrations/2016_06_01_022230_create_blog_category_table.php
@@ -15,6 +15,7 @@ class CreateBlogCategoryTable extends Migration
         Schema::create('blog_category', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title', 120);
+            $table->integer('order')->default(0);
             $table->integer('status')->default(1);
             $table->softDeletes();
             $table->timestamps();