+++ database/migrations/2016_06_02_001633_create_pages_categorys_table.php
@@ -15,6 +15,7 @@ class CreatePagesCategorysTable extends Migration
         Schema::create('pages_categorys', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title', 120);
+            $table->integer('order')->default(0);
             $table->integer('status')->default(1);
             $table->softDeletes();
             $table->timestamps();