+++ database/migrations/2016_06_02_001721_create_pages_contents_table.php
@@ -15,10 +15,12 @@ class CreatePagesContentsTable extends Migration
         Schema::create('pages_contents', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title', 170);
-            $table->text('content');
+            $table->text('description');
             $table->integer('category_id')->unsigned();
             $table->foreign('category_id')->references('id')->on('pages_categorys');
+            $table->integer('order')->default(0);
             $table->integer('status')->default(1);
+            $table->string('slug')->default('');
             $table->string('seo_title', 70);
             $table->string('seo_description', 170);
             $table->softDeletes();