+++ database/migrations/2016_06_08_171056_create_products_category_table.php
@@ -15,6 +15,7 @@ class CreateProductsCategoryTable extends Migration
         Schema::create('products_categorys', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title', 120);
+            $table->integer('order')->default(0);
             $table->integer('status')->default(1);
             $table->softDeletes();
             $table->timestamps();