+++ resources/views/admin/blog/comments/reproved.blade.php
@@ -23,7 +23,7 @@
                                 @lang('admin/blog.posts.edit.posted',['name' => $comment->name, 'title' => $comment->post->title]) <br>
                                 <small class="text-muted">{{ $comment->created_at->diffForHumans() }}</small>
                                 <div class="well">
-                                    {{ $comment->content }}
+                                    {{ $comment->description }}
                                 </div>
                             </div>
                         </div>