+++ resources/views/admin/pages/contents/edit.blade.php
@@ -41,7 +41,7 @@
                             <div class="form-group">
                                 <label class="col-sm-2 control-label">@lang('admin/_globals.forms.content'):</label>
                                 <div class="col-sm-10">
-                                    <textarea name="content" class="form-control summernote">{{ $content->content }}</textarea>
+                                    <textarea name="description" class="form-control summernote">{{ $content->description }}</textarea>
                                 </div>
                             </div>
                             <div class="form-group">